requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-tasks
motor
redis
python-dotenv
